update client san phm

+++ back_end/src/main/java/com/fshoes/core/admin/voucher/model/request/AdCallVoucherOfSell.java
@@ -4,10 +4,12 @@ import com.fshoes.core.common.PageableRequest;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class AdCallVoucherOfSell extends PageableRequest {
     private String idCustomer;
-    private Integer condition;
+    private BigDecimal condition;
     private String textSearch;
 }
